@@ -2,76 +2,88 @@
     "type": "project",
     "license": "proprietary",
     "require": {
-        "php": "^7.4",
+        "php": "^8.1",
         "ext-iconv": "*",
         "ext-json": "*",
-        "calderacc/geobasic": "^0.2.0",
         "calderacc/yourls-api-manager": "^0.2.0",
-        "doctrine/doctrine-migrations-bundle": "^1.0",
+        "ext-simplexml": "*",
+        "composer-plugin-api": "^2.2",
+        "calderacc/geobasic": "^0.5",
+        "composer/package-versions-deprecated": "1.11.99.4",
+        "doctrine/annotations": "^1.0",
+        "doctrine/doctrine-bundle": "^2.5",
+        "doctrine/doctrine-migrations-bundle": "^3.1",
+        "doctrine/orm": "^2.10",
         "dragonmantank/cron-expression": "^2.1",
-        "flagception/flagception-bundle": "^3.2",
+        "flagception/flagception-bundle": "^4.1",
         "fresh/doctrine-enum-bundle": "^6.2",
-        "friendsofsymfony/elastica-bundle": "^5.0",
+        "friendsofsymfony/elastica-bundle": "^6.0",
         "friendsofsymfony/jsrouting-bundle": "^2.2",
         "geocoder-php/nominatim-provider": "^5.2",
-        "hwi/oauth-bundle": "^0.6.1",
-        "jms/serializer-bundle": "^2.3",
+        "hwi/oauth-bundle": "^1.4",
+        "jms/serializer-bundle": "^3.0",
         "jublonet/codebird-php": "^3.1",
-        "knplabs/knp-menu-bundle": "^2.2",
+        "knplabs/knp-menu-bundle": "^3.1",
         "league/csv": "^9.1",
         "league/geotools": "^0.8.0",
+        "maltehuebner/impress-bundle": "^0.1.0",
+        "mhujer/breadcrumbs-bundle": "^1.5",
         "nelmio/api-doc-bundle": "^3.1",
-        "php-amqplib/rabbitmq-bundle": "^1.14",
+        "nesbot/carbon": "^2.43",
+        "php-amqplib/rabbitmq-bundle": "^2.11",
         "php-curl-class/php-curl-class": "^8.0",
-        "php-http/guzzle6-adapter": "^1.1",
-        "php-http/httplug-bundle": "^1.9",
+        "php-http/guzzle6-adapter": "^2.0.2",
+        "php-http/httplug-bundle": "^1.20.0",
         "php-http/message": "^1.7",
+        "phpdocumentor/reflection-docblock": "^5.3",
+        "phpstan/phpdoc-parser": "^1.2",
         "predis/predis": "^1.1",
         "presta/sitemap-bundle": "^1.5",
-        "salavert/time-ago-in-words": "^1.7",
         "sensio/framework-extra-bundle": "^5.1",
-        "sonata-project/admin-bundle": "^3.40",
-        "sonata-project/datagrid-bundle": "^3.2",
-        "sonata-project/doctrine-orm-admin-bundle": "^3.6",
-        "sonata-project/intl-bundle": "^2.5",
         "sonata-project/seo-bundle": "^2.5",
         "symfony/apache-pack": "^1.0",
         "symfony/asset": "^4.0",
+        "symfony/cache": "^5.2",
         "symfony/console": "^4.0",
-        "symfony/debug-pack": "*",
         "symfony/expression-language": "^4.0",
-        "symfony/flex": "^1.0",
+        "symfony/flex": "^2.0",
         "symfony/form": "^4.0",
         "symfony/framework-bundle": "^4.0",
         "symfony/monolog-bundle": "^3.1",
-        "symfony/orm-pack": "*",
         "symfony/process": "^4.0",
+        "symfony/property-access": "^4.0",
+        "symfony/property-info": "^4.0",
+        "symfony/proxy-manager-bridge": "^4.0",
         "symfony/security-bundle": "^4.0",
-        "symfony/serializer-pack": "*",
+        "symfony/serializer": "^4.0",
         "symfony/swiftmailer-bundle": "^3.1",
         "symfony/validator": "^4.0",
         "symfony/web-link": "^4.0",
         "symfony/webpack-encore-bundle": "^1.7",
         "symfony/yaml": "^4.0",
-        "whiteoctober/breadcrumbs-bundle": "^1.4",
-        "willdurand/geocoder-bundle": "^5.0",
-        "ext-simplexml": "*"
+        "willdurand/geocoder-bundle": "^5.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.0",
+        "phpunit/phpunit": "^9.0",
         "symfony/browser-kit": "^4.0",
         "symfony/css-selector": "^4.0",
+        "symfony/debug-bundle": "^4.0",
         "symfony/dotenv": "^4.0",
         "symfony/maker-bundle": "^1.0",
         "symfony/phpunit-bridge": "^4.3",
-        "symfony/profiler-pack": "*",
-        "symfony/test-pack": "^1.0"
+        "symfony/stopwatch": "^4.0",
+        "symfony/web-profiler-bundle": "^4.0"
     },
     "config": {
         "preferred-install": {
             "*": "dist"
         },
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "composer/package-versions-deprecated": true,
+            "composer/installers": true,
+            "symfony/flex": true
+        }
     },
     "autoload": {
         "psr-4": {
