@@ -3,14 +3,14 @@
 namespace App\Analysis\FireworksAnalysis;
 
 use App\Entity\Data;
+use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Bridge\Doctrine\RegistryInterface;
 
 class FireworksModelFactory implements FireworksModelFactoryInterface
 {
-    /** @var RegistryInterface $registry */
-    protected $registry;
+    protected ManagerRegistry $registry;
 
-    public function __construct(RegistryInterface $registry)
+    public function __construct(ManagerRegistry $registry)
     {
         $this->registry = $registry;
     }
@@ -27,22 +27,6 @@ class FireworksModelFactory implements FireworksModelFactoryInterface
                 $resultList[$stationId] = new FireworksModel($data->getStation(), $data, 0);
             }
         }
-
-        return $this->sortResultList($resultList);
-    }
-
-    /**
-     * @TODO this should be done in elasticsearch
-     */
-    protected function sortResultList(array $resultList = []): array
-    {
-        usort($resultList, function(FireworksModel $a, FireworksModel $b): int
-        {
-            if ($a->getData()->getDateTime() === $b->getData()->getDateTime()) {
-                return 0;
-            }
-            return ($a->getData()->getDateTime() > $b->getData()->getDateTime()) ? -1 : 1;
-        });
 
         return $resultList;
     }
