@@ -2,20 +2,15 @@
 
 namespace App\Command\ImportCache;
 
-use App\Entity\Data;
-use App\Pollution\UniqueStrategy\CacheUniqueStrategy;
-use Symfony\Bridge\Doctrine\RegistryInterface;
+use App\Pollution\UniqueStrategy\UniqueStrategyInterface;
 use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
 class ClearImportCacheCommand extends Command
 {
-    /** @var CacheUniqueStrategy $cacheUniqueStrategy */
-    protected $cacheUniqueStrategy;
+    protected UniqueStrategyInterface $cacheUniqueStrategy;
 
     protected static $defaultName = 'luft:import-cache:clear';
 
@@ -24,11 +19,11 @@ class ClearImportCacheCommand extends Command
         $this->setDescription('Add a short description for your command');
     }
 
-    public function __construct(?string $name = null, CacheUniqueStrategy $cacheUniqueStrategy, RegistryInterface $registry)
+    public function __construct(UniqueStrategyInterface $cacheUniqueStrategy)
     {
         $this->cacheUniqueStrategy = $cacheUniqueStrategy->init([]);
 
-        parent::__construct($name);
+        parent::__construct();
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): void
