@@ -3,8 +3,8 @@
 namespace App\Command\ImportCache;
 
 use App\Entity\Data;
-use App\Pollution\UniqueStrategy\CacheUniqueStrategy;
-use Symfony\Bridge\Doctrine\RegistryInterface;
+use App\Pollution\UniqueStrategy\UniqueStrategyInterface;
+use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
@@ -13,11 +13,8 @@ use Symfony\Component\Console\Style\SymfonyStyle;
 
 class PopulateImportCacheCommand extends Command
 {
-    /** @var CacheUniqueStrategy $cacheUniqueStrategy */
-    protected $cacheUniqueStrategy;
-
-    /** @var RegistryInterface $registry */
-    protected $registry;
+    protected UniqueStrategyInterface $cacheUniqueStrategy;
+    protected ManagerRegistry $registry;
 
     protected static $defaultName = 'luft:import-cache:populate';
 
@@ -28,12 +25,12 @@ class PopulateImportCacheCommand extends Command
             ->addOption('interval', 'i', InputOption::VALUE_REQUIRED, 'Provide an interval starting from today backwards', 'P3D');
     }
 
-    public function __construct(?string $name = null, CacheUniqueStrategy $cacheUniqueStrategy, RegistryInterface $registry)
+    public function __construct(UniqueStrategyInterface $cacheUniqueStrategy, ManagerRegistry $registry)
     {
         $this->cacheUniqueStrategy = $cacheUniqueStrategy->init([]);
         $this->registry = $registry;
 
-        parent::__construct($name);
+        parent::__construct();
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): void
