@@ -11,14 +11,13 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 class LoadCitiesCommand extends Command
 {
-    /** @var EntityManagerInterface */
-    protected $entityManager;
+    protected EntityManagerInterface $entityManager;
 
-    public function __construct(?string $name = null, EntityManagerInterface $entityManager)
+    public function __construct(EntityManagerInterface $entityManager)
     {
         $this->entityManager = $entityManager;
 
-        parent::__construct($name);
+        parent::__construct();
     }
 
     protected function configure()
