@@ -17,35 +17,40 @@ class Data
      * @ORM\Column(type="integer")
      * @ORM\GeneratedValue(strategy="AUTO")
      */
-    protected $id;
+    protected ?int $id = null;
 
     /**
-     * @var Station $station
      * @ORM\ManyToOne(targetEntity="Station", inversedBy="datas")
      * @ORM\JoinColumn(name="station_id", referencedColumnName="id")
      * @JMS\Expose()
      * @JMS\Type("App\Entity\Station")
      */
-    protected $station;
+    protected ?Station $station = null;
 
     /**
      * @ORM\Column(type="datetime", nullable=false)
      * @JMS\Expose()
      * @JMS\Type("DateTime<'U'>")
      */
-    protected $dateTime;
+    protected ?\DateTime $dateTime = null;
 
     /**
      * @ORM\Column(type="float", nullable=false)
      * @JMS\Expose()
      */
-    protected $value;
+    protected ?float $value = null;
 
     /**
      * @ORM\Column(type="smallint", nullable=false)
      * @JMS\Expose()
      */
-    protected $pollutant;
+    protected ?int $pollutant = null;
+
+    /**
+     * @ORM\Column(type="string", length=32, nullable=true)
+     * @JMS\Expose()
+     */
+    private $tag;
 
     public function setId(int $id): Data
     {
@@ -59,7 +64,7 @@ class Data
         return $this->id;
     }
 
-    public function getStation(): Station
+    public function getStation(): ?Station
     {
         return $this->station;
     }
@@ -97,7 +102,7 @@ class Data
         return $this;
     }
 
-    public function getValue(): float
+    public function getValue(): ?float
     {
         return $this->value;
     }
@@ -109,7 +114,7 @@ class Data
         return $this;
     }
 
-    public function getPollutant(): int
+    public function getPollutant(): ?int
     {
         return $this->pollutant;
     }
@@ -121,7 +126,7 @@ class Data
         return $this;
     }
 
-    public function getProvider(): string
+    public function getProvider(): ?string
     {
         return $this->station->getProvider();
     }
@@ -132,5 +137,17 @@ class Data
         $dateTime->sub(new \DateInterval('P1W'));
 
         return $dateTime >= $this->dateTime;
+    }
+
+    public function getTag(): ?string
+    {
+        return $this->tag;
+    }
+
+    public function setTag(?string $tag): self
+    {
+        $this->tag = $tag;
+
+        return $this;
     }
 }
