@@ -7,24 +7,20 @@ use App\Entity\Station;
 use App\Pollution\DataCache\DataCacheInterface;
 use App\Pollution\DataCache\KeyGenerator;
 use Caldera\GeoBasic\Coord\CoordInterface;
-use function PHPSTORM_META\elementType;
-use Symfony\Bridge\Doctrine\RegistryInterface;
+use Doctrine\Persistence\ManagerRegistry;
 
 class Co2CachedDataRetriever implements DataRetrieverInterface
 {
-    /** @var DataCacheInterface $dataCache */
-    protected $dataCache;
+    protected DataCacheInterface $dataCache;
+    protected ManagerRegistry $registry;
 
-    /** @var RegistryInterface $registry */
-    protected $registry;
-
-    public function __construct(DataCacheInterface $dataCache, RegistryInterface $registry)
+    public function __construct(DataCacheInterface $dataCache, ManagerRegistry $registry)
     {
         $this->dataCache = $dataCache;
         $this->registry = $registry;
     }
 
-    public function retrieveDataForCoord(CoordInterface $coord, int $pollutantId, \DateTime $fromDateTime = null, \DateInterval $dateInterval = null, float $maxDistance = 20.0, int $maxResults = 750): array
+    public function retrieveDataForCoord(CoordInterface $coord, int $pollutantId = null, \DateTime $fromDateTime = null, \DateInterval $dateInterval = null, float $maxDistance = 20.0, int $maxResults = 750): array
     {
         if ($coord instanceof Station) { // TODO this should be regulated through a service
             return [];
