@@ -3,75 +3,99 @@
 namespace App\Pollution\DataRetriever;
 
 use App\Entity\Station;
+use App\Pollution\DataFinder\ElasticFinder;
+use App\Pollution\StationCache\StationCacheInterface;
 use Caldera\GeoBasic\Coord\CoordInterface;
-use FOS\ElasticaBundle\Finder\FinderInterface;
 
 class ElasticDataRetriever implements DataRetrieverInterface
 {
-    /** @var FinderInterface $dataFinder */
-    protected $dataFinder;
+    protected StationCacheInterface $stationCache;
 
-    public function __construct(FinderInterface $dataFinder)
+    protected ElasticFinder $finder;
+
+    public function __construct(ElasticFinder $finder, StationCacheInterface $stationCache)
     {
-        $this->dataFinder = $dataFinder;
+        $this->finder = $finder;
+        $this->stationCache = $stationCache;
     }
 
-    public function retrieveDataForCoord(CoordInterface $coord, int $pollutantId, \DateTime $fromDateTime = null, \DateInterval $dateInterval = null, float $maxDistance = 20.0, int $maxResults = 750): array
+    public function retrieveDataForCoord(CoordInterface $coord, int $pollutantId = null, \DateTime $fromDateTime = null, \DateInterval $dateInterval = null, float $maxDistance = 20.0, int $maxResults = 750): array
     {
-        if ($coord instanceof Station) {
-            $stationQuery = new \Elastica\Query\Nested();
-            $stationQuery->setPath('station');
-            $stationQuery->setQuery(new \Elastica\Query\Term(['station.id' => $coord->getId()]));
-        } else {
-            $stationGeoQuery = new \Elastica\Query\GeoDistance('station.pin', [
-                'lat' => $coord->getLatitude(),
-                'lon' => $coord->getLongitude(),
-            ],
-                sprintf('%fkm', $maxDistance));
+        $fromDateTime = new \DateTime();
+        $fromDateTime->sub(new \DateInterval('PT8H'));
 
-            $stationQuery = new \Elastica\Query\Nested();
-            $stationQuery->setPath('station');
-            $stationQuery->setQuery($stationGeoQuery);
-        }
-
-        $pollutantQuery = new \Elastica\Query\Term(['pollutant' => $pollutantId]);
-
-        $boolQuery = new \Elastica\Query\BoolQuery();
-        $boolQuery
-            ->addMust($pollutantQuery)
-            ->addMust($stationQuery);
+        $query = new \Elastica\Query(new \Elastica\Query\MatchAll());
+        $query->setSize(0);
 
         if ($fromDateTime && $dateInterval) {
-            $untilDateTime = clone $fromDateTime;
-            $untilDateTime->add($dateInterval);
+            $untilDateTime = (clone $fromDateTime)->add($dateInterval);
+            $untilDateTime = new \DateTime();
 
-            $dateTimeQuery = new \Elastica\Query\Range('dateTime', [
-                'gt' => $fromDateTime->format('Y-m-d H:i:s'),
-                'lte' => $untilDateTime->format('Y-m-d H:i:s'),
-                'format' => 'yyyy-MM-dd HH:mm:ss',
-            ]);
+            $dateTimeAggregation = new \Elastica\Aggregation\Range('datetime_agg');
+            $dateTimeAggregation->setField('dateTime');
+            $dateTimeAggregation->addRange(
+                $fromDateTime->format('Y-m-d H:i:s'),
+                $untilDateTime->format('Y-m-d H:i:s')
+            );
 
-            $boolQuery->addMust($dateTimeQuery);
+            $query->addAggregation($dateTimeAggregation);
         }
 
-        $query = new \Elastica\Query($boolQuery);
+        $pollutantAggregation = new \Elastica\Aggregation\Terms('pollutant_agg');
+        $pollutantAggregation->setField('pollutant');
 
-        $query
-            ->addSort([
-                '_geo_distance' => [
-                    'station.pin' => [
-                        'lat' => $coord->getLatitude(),
-                        'lon' => $coord->getLongitude()
-                    ],
-                    'order' => 'asc',
-                    'unit' => 'km',
-                    'nested_path' => 'station',
-                ]
-            ])
-            ->addSort(['dateTime' => 'desc']);
+        if ($fromDateTime && $dateInterval) {
+            $dateTimeAggregation->addAggregation($pollutantAggregation);
+        } else {
+            $query->addAggregation($pollutantAggregation);
+        }
 
-        $query->setSize($maxResults);
+        if ($coord instanceof Station) {
+            $stationAggregation = new \Elastica\Aggregation\Terms('station_agg');
+            $stationAggregation->setField('stationCode');
+            $stationAggregation->setInclude($coord->getStationCode());
 
-        return $this->dataFinder->find($query);
+            $pollutantAggregation->addAggregation($stationAggregation);
+        } else {
+            $providerAggregation = new \Elastica\Aggregation\Terms('provider_agg');
+            $providerAggregation->setField('provider');
+            $pollutantAggregation->addAggregation($providerAggregation);
+
+            $geodistanceAggregation = new \Elastica\Aggregation\GeoDistance(
+                'geodistance_agg',
+                'pin',
+                $coord->toLatLonArray()
+            );
+            $geodistanceAggregation->addRange(null, $maxDistance);
+            $geodistanceAggregation->setUnit('km');
+
+            $providerAggregation->addAggregation($geodistanceAggregation);
+        }
+
+        $topHitsAggregation = new \Elastica\Aggregation\TopHits('top_hits_agg');
+        $topHitsAggregation->setSize(1);
+        $topHitsAggregation->setSort([[
+            '_geo_distance' => [
+                'pin' => [
+                    'lat' => $coord->getLatitude(),
+                    'lon' => $coord->getLongitude(),
+                ],
+                'order' => 'ASC',
+                'unit' => 'km',
+            ],
+            'dateTime' => 'DESC',
+        ],
+        ]);
+
+        if ($coord instanceof Station) {
+            $stationAggregation->addAggregation($topHitsAggregation);
+        } else {
+            $geodistanceAggregation->addAggregation($topHitsAggregation);
+        }
+
+        //dd(json_encode($query->toArray()));
+        $result = $this->finder->find($query);
+
+        return $result;
     }
 }
