@@ -4,17 +4,14 @@ namespace App\Pollution\UniqueStrategy;
 
 use App\Entity\Data;
 use App\Pollution\Value\Value;
-use Symfony\Bridge\Doctrine\RegistryInterface;
+use Doctrine\Persistence\ManagerRegistry;
 
 class OrmUniqueStrategy implements UniqueStrategyInterface
 {
-    /** @var RegistryInterface $registry */
-    protected $registry;
+    protected ManagerRegistry $registry;
+    protected array $existentDataList = [];
 
-    /** @var array $existentDataList */
-    protected $existentDataList = [];
-
-    public function __construct(RegistryInterface $registry)
+    public function __construct(ManagerRegistry $registry)
     {
         $this->registry = $registry;
     }
