@@ -5,8 +5,10 @@ namespace App\Twig\Extension;
 use App\Air\ViewModel\MeasurementViewModel;
 use App\Air\AirQuality\Calculator\AirQualityCalculatorInterface;
 use App\Air\AirQuality\PollutionLevel\PollutionLevelInterface;
+use Twig\Extension\AbstractExtension;
+use Twig\TwigFunction;
 
-class PollutionLevelTwigExtension extends \Twig_Extension
+class PollutionLevelTwigExtension extends AbstractExtension
 {
     protected $backgroundColors = [
         0 => 'white',
@@ -28,8 +30,7 @@ class PollutionLevelTwigExtension extends \Twig_Extension
         6 => 'red',
     ];
 
-    /** @var AirQualityCalculatorInterface $airQualityCalculator */
-    protected $airQualityCalculator;
+    protected AirQualityCalculatorInterface $airQualityCalculator;
 
     public function __construct(AirQualityCalculatorInterface $airQualityCalculator)
     {
@@ -39,10 +40,10 @@ class PollutionLevelTwigExtension extends \Twig_Extension
     public function getFunctions(): array
     {
         return [
-            new \Twig_SimpleFunction('max_pollution_level', [$this, 'maxPollutionLevel'], ['is_safe' => ['raw']]),
-            new \Twig_SimpleFunction('pollution_color', [$this, 'pollutionColor'], ['is_safe' => ['raw']]),
-            new \Twig_SimpleFunction('pollution_color_name', [$this, 'pollutionColorName'], ['is_safe' => ['raw']]),
-            new \Twig_SimpleFunction('pollution_levels', [$this, 'getLevelsForMeasurement'], ['is_safe' => ['raw']]),
+            new TwigFunction('max_pollution_level', [$this, 'maxPollutionLevel'], ['is_safe' => ['raw']]),
+            new TwigFunction('pollution_color', [$this, 'pollutionColor'], ['is_safe' => ['raw']]),
+            new TwigFunction('pollution_color_name', [$this, 'pollutionColorName'], ['is_safe' => ['raw']]),
+            new TwigFunction('pollution_levels', [$this, 'getLevelsForMeasurement'], ['is_safe' => ['raw']]),
         ];
     }
 
@@ -85,4 +86,3 @@ class PollutionLevelTwigExtension extends \Twig_Extension
         return 'pollution_level_extension';
     }
 }
-
