@@ -2,6 +2,7 @@ import 'bootstrap';
 import Geolocation from './modules/Geolocation';
 import Map from './modules/Map';
 import Search from './modules/Search';
+import TimeAgoInWords from './modules/TimeAgoInWords';
 import History from './modules/History';
 
-export {Geolocation, Map, Search, History}
+export {Geolocation, Map, Search, History, TimeAgoInWords}
