@@ -27,9 +27,14 @@ class FireworksAnalysis extends AbstractFireworksAnalysis
             ->addMust($pollutantQuery)
             ->addMust($dateTimeQuery)
             ->addMust($pollutionQuery)
-            ->addMust($providerQuery);
+            //->addMust($providerQuery)
+        ;
 
         $query = new \Elastica\Query($boolQuery);
+        $query
+                ->addSort(['value' => 'desc'])
+            ->addSort(['dateTime' => 'desc'])
+        ;
 
         $results = $this->finder->find($query, 5000);
 
