@@ -2,7 +2,6 @@
 
 namespace App\DataPurger;
 
-use App\Entity\Data;
 use App\Provider\ProviderInterface;
 use Doctrine\Persistence\ManagerRegistry;
 use Elastica\Client;
@@ -19,21 +18,35 @@ class FastDataPurger implements DataPurgerInterface
         $this->client = $client;
     }
 
-    public function purgeData(\DateTimeInterface $untilDateTime, ProviderInterface $provider = null): int
+    public function purge(\DateTime $untilDateTime, ProviderInterface $provider = null, bool $withTags): int
     {
-        $counter = $this->purgeDatabase($untilDateTime, $provider);
-
-        $this->purgeElasticsearch($untilDateTime, $provider);
+        $counter = $this->countData($untilDateTime, $provider, $withTags);
+        $this->purgeData($untilDateTime, $provider, $withTags);
 
         return $counter;
     }
 
-    public function purgeDatabase(\DateTimeInterface $untilDateTime, ProviderInterface $provider = null): int
+    public function countData(\DateTime $untilDateTime, ProviderInterface $provider = null, bool $withTags): int
     {
-        return $this->managerRegistry->getRepository(Data::class)->deleteData($untilDateTime, $provider);
+        $query = $this->buildQuery($untilDateTime, $provider, $withTags);
+
+        $result = $this->client->request('air_data/_count', Request::GET, $query);
+
+        $response = $result->getData();
+
+        $count = $response['count'];
+
+        return $count;
     }
 
-    public function purgeElasticsearch(\DateTimeInterface $untilDateTime, ProviderInterface $provider = null): int
+    public function purgeData(\DateTime $untilDateTime, ProviderInterface $provider = null, bool $withTags): void
+    {
+        $query = $this->buildQuery($untilDateTime, $provider, $withTags);
+
+        $this->client->request('air_data/_delete_by_query', Request::POST, $query);
+    }
+
+    protected function buildQuery(\DateTime $untilDateTime, ProviderInterface $provider = null, bool $withTags): array
     {
         $query = [
             'query' => [
@@ -58,8 +71,10 @@ class FastDataPurger implements DataPurgerInterface
             ];
         }
 
-        $this->client->request('air_data/_delete_by_query', Request::POST, $query);
+        if (!$withTags) {
+            $query['query']['bool']['must_not'][]['exists'] = ['field' => 'tag'];
+        }
 
-        return 0;
+        return $query;
     }
 }
