@@ -210,7 +210,7 @@ class City
         return $this->fahrverboteSlug !== null;
     }
 
-    public function __toString(): ?string
+    public function __toString(): string
     {
         return $this->name ? $this->name : '';
     }
