@@ -5,6 +5,7 @@ namespace App\Geocoding\Guesser;
 use Caldera\GeoBasic\Coord\Coord;
 use Geocoder\Model\AddressCollection;
 use Geocoder\Query\ReverseQuery;
+use Http\Client\Exception\NetworkException;
 
 class CityGuesser extends AbstractGuesser implements CityGuesserInterface
 {
@@ -19,8 +20,12 @@ class CityGuesser extends AbstractGuesser implements CityGuesserInterface
     {
         $query = ReverseQuery::fromCoordinates($coord->getLatitude(), $coord->getLongitude());
 
-        /** @var AddressCollection $addressCollection */
-        $addressCollection = $this->provider->reverseQuery($query);
+        try {
+            /** @var AddressCollection $addressCollection */
+            $addressCollection = $this->provider->reverseQuery($query);
+        } catch (NetworkException $networkException) {
+            return null;
+        }
 
         if (!$addressCollection || !$addressCollection->first() || (!$addressCollection->first()->getLocality() && !$addressCollection->first()->getSubLocality())) {
             return null;
