@@ -5,22 +5,24 @@ namespace App\Geocoding\RequestConverter;
 use App\Entity\Zip;
 use App\Geocoding\Geocoder\GeocoderInterface;
 use Caldera\GeoBasic\Coord\Coord;
-use Symfony\Bridge\Doctrine\RegistryInterface;
+use Caldera\GeoBasic\Coordinate\Coordinate;
+use Caldera\GeoBasic\Coordinate\CoordinateInterface;
+use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Component\HttpFoundation\Request;
 
 class RequestConverter implements RequestConverterInterface
 {
     protected GeocoderInterface $geocoder;
 
-    protected RegistryInterface $registry;
+    protected ManagerRegistry $registry;
 
-    public function __construct(GeocoderInterface $geocoder, RegistryInterface $registry)
+    public function __construct(GeocoderInterface $geocoder, ManagerRegistry $registry)
     {
         $this->geocoder = $geocoder;
         $this->registry = $registry;
     }
 
-    public function getCoordByRequest(Request $request): ?Coord
+    public function getCoordByRequest(Request $request): ?CoordinateInterface
     {
         $latitude = (float) $request->query->get('latitude');
         $longitude = (float) $request->query->get('longitude');
@@ -34,7 +36,7 @@ class RequestConverter implements RequestConverterInterface
         }
 
         if ($latitude && $longitude) {
-            $coord = new Coord(
+            $coord = new Coordinate(
                 $latitude,
                 $longitude
             );
@@ -48,7 +50,7 @@ class RequestConverter implements RequestConverterInterface
             $firstResult = array_pop($result);
 
             if ($firstResult) {
-                $coord = new Coord($firstResult['value']['latitude'], $firstResult['value']['longitude']);
+                $coord = new Coordinate($firstResult['value']['latitude'], $firstResult['value']['longitude']);
 
                 return $coord;
             }
