@@ -1,4 +1,5 @@
-<?php declare(strict_types=1);
+<?php
+declare(strict_types=1);
 
 namespace App\Pollution\DataPersister;
 
@@ -40,7 +41,7 @@ class CachePersister extends AbstractPersister
 
             $this->dataCache->addData($data);
         }
-        
+
         return $this;
     }
 
